flask-mongodb/requirements.txt
flask
pymongo
redis
flask_caching
python-dotenv
pytest 
requests